+++ libs/Value.hpp
@@ -35,7 +35,7 @@ private:
 public:
   HP() = default;
   HP(int val);
-  void takeDamage(int damage);
+  void takeDamage(double damage);
   void heal(int amount);
 };
 // Derived class for mana points
